Moved pages to pages folder
