+++ src/App.js
@@ -3,20 +3,20 @@ import './App.css';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
 import { Route, BrowserRouter as Router, Switch } from 'react-router-dom';
 
-import AboutPage from './component/AboutPage';
-import AddMemoryPage from './component/AddMemoryPage';
-import BondRequestsPage from './component/BondRequestsPage';
-import LandingPage from './component/LandingPage';
+import AboutPage from './pages/AboutPage';
+import AddMemoryPage from './pages/AddMemoryPage';
+import BondRequestsPage from './pages/BondRequestsPage';
+import LandingPage from './pages/LandingPage';
 import LoadingOverlay from 'react-loading-overlay';
-import LoginPage from './component/LoginPage';
-import MemolinePage from './component/MemolinePage';
-import MyPeoplePage from './component/MyPeoplePage';
+import LoginPage from './pages/LoginPage';
+import MemolinePage from './pages/MemolinePage';
+import MyPeoplePage from './pages/MyPeoplePage';
 import NavBar from './component/NavBar/NavBar';
-import PageNotFound from './component/404';
-import ProfilePage from './component/ProfilePage/ProfilePage';
+import PageNotFound from './pages/404';
+import ProfilePage from './pages/ProfilePage/ProfilePage';
 import React from 'react';
-import RegisterPage from './component/RegisterPage';
-import SearchPage from './component/SearchPage';
+import RegisterPage from './pages/RegisterPage';
+import SearchPage from './pages/SearchPage';
 import SessionLogin from './component/SessionLogin/SessionLogin';
 import { connect } from 'react-redux';
 import { isLoading } from './selectors';
