+++ src/pages/RegisterPage/RegisterPage.jsx
@@ -5,9 +5,9 @@ import { toastError, toastSuccess } from "../../utils/Toast";
 import ApiService from "../../service/ApiService";
 import Card from "react-bootstrap/Card";
 import { Form } from "react-bootstrap";
-import Input from "../Form/Input";
+import Input from "../../component/Form/Input";
 import React from "react";
-import RedirectIf from "../RedirectIf";
+import RedirectIf from "../../component/RedirectIf";
 import checkUsername from "../../utils/Validation";
 import { connect } from "react-redux";
 import { getUserState } from "../../selectors";
