+++ src/pages/RegisterPage/package.json
