+++ src/pages/SearchPage/SearchPage.css
